Check conditions on delete => ASSETS

+++ app/Controllers/AssetCategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Controllers;
 
 use App\Controllers\Repositories\AssetCategoryRepository;
+use App\Controllers\Repositories\AssetRepository;
 use App\Libraries\ServerSideDataTable;
 use CodeIgniter\API\ResponseTrait;
 
@@ -12,10 +13,12 @@ class AssetCategoryController extends BaseController
     use ResponseTrait;
 
     private $AssetCategoryRepository;
+    private $AssetRepository;
 
     public function __construct()
     {
         $this->AssetCategoryRepository = new AssetCategoryRepository();
+        $this->AssetRepository = new AssetRepository();
     }
 
     public function assetCategory()
@@ -61,6 +64,10 @@ class AssetCategoryController extends BaseController
     public function delete()
     {
         $id = $this->request->getPost('id');
+
+        $check = $this->AssetRepository->assetByCategories([$id]);
+        if(!empty($check))
+            return $this->respond(responseJson(202, true, ['mag' => 'This category cannot be deleted because there are assets associated with it.']));
 
         $result = $this->AssetCategoryRepository->deleteAssetCategory($id);
         $result = $result
